Create abstract class that *Creator test files will inherit. Rename RoomProcessorTest -> EntityFilesCreatorTest to be more specific. Create file for RoomDataCreator tests.
