+++ androidutilities-room/src/test/kotlin/com/heyzeusv/androidutitlities/room/EntityFilesCreatorTest.kt
@@ -1,16 +1,9 @@
 package com.heyzeusv.androidutitlities.room
 
-import com.heyzeusv.androidutilities.room.RoomProcessorProvider
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
-import com.tschuchort.compiletesting.kspIncremental
-import com.tschuchort.compiletesting.kspSourcesDir
-import com.tschuchort.compiletesting.symbolProcessorProviders
 import org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi
-import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.TemporaryFolder
-import java.io.File
 import kotlin.test.assertEquals
 
 /**
@@ -18,10 +11,7 @@ import kotlin.test.assertEquals
  *  as a guide to write these tests.
  */
 @OptIn(ExperimentalCompilerApi::class)
-class RoomProcessorTest  {
-
-    @get:Rule
-    val tempFolder: TemporaryFolder = TemporaryFolder()
+class EntityFilesCreatorTest : CreatorTestBase()  {
 
     @Test
     fun `Generate basic entity with two fields`() {
@@ -393,39 +383,6 @@ class RoomProcessorTest  {
             assertEquals(expectedEntityWithEveryAnnotation, generatedFileText)
         }
     }
-
-    private fun compile(vararg sourceFiles: SourceFile): KspCompileResult {
-        val compilation = prepareCompilation(*sourceFiles)
-        val result = compilation.compile()
-        return KspCompileResult(
-            result = result,
-            generatedFiles = findGeneratedFiles(compilation)
-        )
-    }
-
-    private fun prepareCompilation(vararg sourceFiles: SourceFile): KotlinCompilation =
-        KotlinCompilation()
-            .apply {
-                workingDir = tempFolder.root
-                inheritClassPath = true
-                symbolProcessorProviders = listOf(RoomProcessorProvider())
-                sources = sourceFiles.asList()
-                verbose = false
-                kspIncremental = true
-                messageOutputStream = System.out
-            }
-
-    private fun findGeneratedFiles(compilation: KotlinCompilation): List<File> {
-        return compilation.kspSourcesDir
-            .walkTopDown()
-            .filter { it.isFile }
-            .toList()
-    }
-
-    private data class KspCompileResult(
-        val result: KotlinCompilation.Result,
-        val generatedFiles: List<File>,
-    )
 
     companion object {
         private fun basicEntityTwoFields(name: String, tableName: String = name) = """
