+++ androidutilities-room/src/test/kotlin/com/heyzeusv/androidutitlities/room/RoomDataCreatorTest.kt
@@ -1,0 +1,5 @@
+package com.heyzeusv.androidutitlities.room
+
+class RoomDataCreatorTest : CreatorTestBase() {
+
+}
